Provide ability to flush and clear the session when using the dynamic task callback. After leaving the Zendesk import running, it was getting slower and slower and at the rate it was going was going to take nearly 2 weeks!
I eventually traced this down to the fact that a transaction was in use. Every single written entity was being kept in memory for the duration of the transaction. In order to combat this, allowing a transaction, and a usable import, I added a couple of methods to the dynamic callback to allow the session to be flushed and cleared (if transactional). The flush and clear currently happens every 10 rows processed by the Xml import tas (maybe this should happen at a higher level?). This keeps the memory usage of the task itself limited to the (not inconsiderable) xml Document instance plus 10 IssueResource instances while still having a database level transaction.

The only thing that should be lost by doing is some cascade functionality during the transaction, which is not used anyway.