Add HTML i18n resources and allow messages to be used without a template.
I needed to be able to specify specific HTML content as a message template. I added support for retrieving from the i18n files when creating MessageResource. It will also use the HTML content with a template (i.e <Not Set>). When operating like this, the actual content from the message is used and it is not wrapped by a template.