#V1HT61 - Customer replies to external ticket with attachment, doesn't appear in ticket
It appears that attaching emails when using MessageResourceService/EmailNotificationService with HTML emails does not work. This is because EmailAttachment is passed a single instance of an InputStream for the attachment data. However, EmailAttachment ultimately implements javax.activation.DataSource. The JavaDocs for this class say

"This method returns an InputStream representing the data and throws the appropriate exception if it can not do so. Note that a new InputStream object must be returned each time this method is called, and the stream must be positioned at the beginning of the data. "

So getInputStream() MUST return a new instance whenever it is called. When JavaMail sends an HTML with attachments, it does indeed call this method multiple times. The first is apparently when email headers are being created, the second to get the content itself. This second call was failing saying the stream was already closed.

Instead of EmailAttachment being passed an InputStream, it is now an abstract class that must implemented getInputStream.