Adds a 'Jobs' page consisting of Local and Clustered jobs, and sits alongside the dashboard and sessions tabs.
While working on the account linking stuff it would have been useful to see what jobs run when without restoring to querying the database directly. I found that somebody had started a 'jobs.html' but that it wasn't actually hooked up to anything. Same for connector synchronizes.

This commit fleshes out that work and allows view, search and removal of jobs by any Administrator or user with SYSTEM permission (Jobs are not currently realm specific and such a change would be too invasive for what I wanted to do here). Searching, sorting on all columns should work.

Internally, this commit basically presents Quartz Triggers as a 'ScheduledResource' from which point on it is more or less treated as a normal resource. ScheduleResource instances are created om the fly and never stored, so there is no repository implementation, and thus the property template is programatically created instead of comming from XML (maybe this could have come from XML, I am not sure if there is a way to do this with such pseudo-resources).

This doesn't quite give us all the features the NAM version of this had (e.g. rescheduling, job progress and cancelling come to mind), but at least the actually active jobs are visible including next/last run times.