Removed RealmService.allUsers() and RealmService.allGroups() and a number of other methods and resources that returns Lists of objects and replaced them with iterations.
@ludup could you please review.

Hibernate does not provide us with a Collection backed by paging, so all results are loaded into memory. For even moderately large realms, this is problematic.

So instead, any time we might need to iterate over all users or groups, or any other large list of objects returned from the database, Collection<SomeResource> Dao methods should be avoid and Iterator<SomeResource> should be used instead.

There are some occasions where this collection may need to be iterated over more than once. In this case, the Iteratable<SomeResource> should be used instead.

This commit contains a number of special Iterator implementations to aid the migration. PagedIterator is the primary one (renamed and re-namespaced for this commit), but there are new ones too. For example, CompoundIterator (which iterates over one or more others Iterators presenting them as a single one), or TransformingIterator which can transform each iterated object, or ProxiedIterator which can be used to build other state-machine like iterators.

TODO: There is still a major sticking point here which would come into when Everyone or other large roles might be used with resource role assignement. In this case, these new iterators are turned back into Sets or Lists or whatever and so negate this fix.