A very simple HTTP/HTTPS server that is started very early in the server initalisation.
This allows administrators to immediately connect and be given feedback on the state of server startup. When the server is completely ready, this HTTP server is shutdown and the full blown Netty server started in its place.

To capture log output on startup, an additional log4j appender must be configured. This needs to be added to the various log4j.properties files.