There have been multiple requests for Hyperticket to show "Ticket Email Events", i.e. whenever an email is sent to a customer as the result of a ticket action (open, comment, close etc), that an event is added to the ticket timeline confirming when the email was actually sent via SMTP.
This became difficult when email batches were introduced, as we lost the association between the email event and the ticket. This would have previously possible using a ThreadLocal, but now we need another way.

To support this, I have added a 'Context' variable to EmailBatchItem, that can be passed all the way down through MessageResource into the final EmailEvent that is sent when the email is actually sent.

In the case of Hyperticket, I supply a context value of "ticket:#QUEUE123". I then listen for EmailEvent and if there is a context starting with ticket: then I can extract the ticket and now know the email has been sent. If this feature is useful for other emails, then they should use their own namespace for the prefix.