LDAP proxy for Linux clients (and other POSIX systems such as OS X etc).
This involved completing more of the LDAP servers features :-

  * Subtree scopes needed to be working properly
  * Attribute filtering needed to be working (there are still some operations that need adding but these will be trivial)

Other changes include :

  * Native attribute names must now be prefixed with a % symbol to differentiate between replacement variables, scripts and plain text.

I wanted the ability to be able to be retrieve a users groups (or a groups' groups), but the interfaces did  not allow this without lots of casting. So I exposed this in UserPrincipal and GroupPrincipal as an abstract method with a generic type parameter. This involved adding the wildcard to a number of instances of where these specialised types were used over the root interface Principal. This actually no reason why Prnicipal itself cannot have getGroups(), as all implementations satisfy this except for FakePrincipal which could easily have a FakeGroup added as well (just for the generics compile time safety).  However, this would involving adding the wildcard to *1000s* of usages, instead of about the 100 or so needed for the more specialised interfaces.

It should be noted that for it to be possible to login to a Linux machine, the user MUST have a group attached, and that group must be exposed in LDAP. We could possibly generate some kind of virtual group if this is not the case, but at the moment it is not a big task to add a group to use for LDAP users.