or the realm deletion work, a new phase in the upgrade system has to be introduce. 'pre.sql'. These scripts are supposed to be run BEFORE hibernate 'takes over' database operations, and does its own schema generation. This was necessary for upgrades to work.  However .....
These pre.sql scripts werent getting run on the first run with a fresh install.  This was because at this stage, the 'upgrade' table doesnt exist at all (and so would error if you tried). Originally, to overcome this, i just said "if pre.sql scripts, and this is a fresh install, then dont do ANYTHING".  I thought this was fine, because the whole reason for the pre scripts was for upgrades

However ....  this meant that on the 2nd boot, because the upgrade row hadnt been written to tell it to not run a particular script, it then runs the pre.sql. This time, the database is not  fresh, so it thinks its an upgrade and drops all foreign keys, and then lets hibernate recreate (incorrectly)

The fix for this was, during FIRST boot, on a FRESH install, also make sure upgrade rows are written for the pre scripts.