#M176
Generic Paged iterator would only iterate the first page if it was using Iterator.remove() as it iterated. This particular pattern was used by the batching system. Due to PagedIterator.remove() incorrectly not being final, it wasn't noticed that AbstractReposityImpl's usage of a generic iterator was overriding the wrong method. This meant the start position was not decreased by one when an item was removed.