Consistent firewall rule handling for network services.
This started as a fix to Wireguard, to ensure it's port is opened on the firewall, but I realised all extensions that add network services suffer the same problem.

There is now an extension to the ManageableService interface, IPManageableService that requires a service provide the ports (TCP or UDP) that it will listen on, and also to implement a simple listener that is notified when the configuration changes.

The x-hypersocket-vm module hooks into these listeners and adjusts firewall configuration on the VM accordingly.