More jackson issues.
Latest identity4j-office365 MUST have Jackson 2.9.x or later.

Have aligned to 2.9.10, except for jackson-datatype-hibernate5 which MUST be kept at version 2.8.11. This is due to Hibernate requirements. To fix this long term, we need to upgrade to Hibernate 5.2 or later. This implies update many other framework jars including Spring and various associated libraries.