M390 - If a wildcard certificate is not validated before restart, service will fail to start up
Certificate providers can now have hook method to answer clients a certificate provided by it is ready for production or not. Typically lets encrypt wild card certs require DNS verification till the time it is not validated, the certificate is not ready, i.e. it is null. In case admin selects it for an interface the server will not restart and will be a burnt toast.