#M736 - Database page incorrectly showing MySQL on a Windows install
It was found that system properties were not getting stored in weight order, logic while saving database was breaking, it was saving port before jdbc vendor leading to incorrect assumption current vendor is default H2 and we are trying to add port.

Added logic to first sort the values by weight and then in order pass it to save.