#M739 - Suspended accounts do not get resumed as expected
It was found that in some scenarios when creating suspension passed in String username was passed and during resume compared to principal name, for most cases it worked fine, but in some cases when invalid credentials were passed and accounts were locked it was using principal email,  as system allows it, but during resume the logic was comparing it to principal name leading to no row found for suspension.

Now all suspension logic is expecting Principal object and suspension name and job id creation is local to class or in methods, no passed in argument is used.