#M768 - Expired license should disable more features
Added new interface MessageDeliveryPreSendCheck which can be implemented to provide message delivery pre send check.

This makes use of Java SPI to locate an implementation as we want to control how this instance is registered with spring.

MessageDeliveryPreSendCheckConfiguration is a simple spring configuration which provides the selected implementation of pre send check. IN case none is provided it defaults to a simple implementation.

Java SPI also used as if not SPring would pick up all registered implementation we need tight control we require only one, makes things simple, but yes configuration needs a little work, as implementations needed to be in src/main/resources/META-INF/services/...

MessageDeliveryBuilder's send method makes use of the provided implementation to check message should be sent or not.