#M787 - Add retry capability to cloud messaging
Added message retry util, which will retry sending in case of failed exception for IOException or MailException.

It returns result in a bean which encapsulates state in which retry ended up,

1. success
2. exception thrown that is of no interest
3. all retry attempts made still o success
4. worst unrecoverable

Client code is responsible to probe this state and act accordingly.

For this new configuration also added to enter number of times to retry and retry in how many seconds.